+++ src/main/java/seedu/toluist/controller/UpdateTaskController.java
@@ -42,7 +42,7 @@ public class UpdateTaskController extends Controller {
         String description = tokens.get(TaskTokenizer.TASK_DESCRIPTION);
 
         String indexToken = tokens.get(TaskTokenizer.TASK_VIEW_INDEX);
-        Task task = todoList.getTask(indexToken);
+        Task task = uiStore.getTask(indexToken);
 
         String startDateToken = tokens.get(TaskTokenizer.TASK_START_DATE_KEYWORD);
         LocalDateTime startDateTime = DateTimeUtil.parseDateString(startDateToken);
