+++ src/main/java/seedu/toluist/ui/UiStore.java
@@ -2,6 +2,7 @@ package seedu.toluist.ui;
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -25,8 +26,22 @@ public class UiStore {
         this.viewedTasks = tasks;
     }
 
+    public Task getTask(String indexToken) {
+        int index = indexToken == null ? -1 : Integer.parseInt(indexToken) - 1;
+        Task task = indexToken == null ? null : viewedTasks.get(index);
+        return task;
+    }
+
     public ArrayList<Task> getTasks() {
         return viewedTasks;
+    }
+
+    public ArrayList<Task> getTasks(List<Integer> indexes) {
+        ArrayList<Task> tasks = new ArrayList<Task>();
+        for (int index: indexes) {
+            tasks.add(viewedTasks.get(index - 1));
+        }
+        return tasks;
     }
 
     public ObservableList<Task> getUiTasks() {
